Copie du staging/production pour tester l'affichage du bouton reset

+++ app/controllers/tools_controller.rb
@@ -2,10 +2,12 @@ class ToolsController < ApplicationController
   skip_before_action :authentifie
 
   def reset_base
-    Evenement.delete_all
-    Occupant.delete_all
-    Invitation.delete_all
-    Projet.delete_all
-    redirect_to root_path, notice: t('reinitialisation.succes')
+    unless Rails.env.production?
+      Evenement.delete_all
+      Occupant.delete_all
+      Invitation.delete_all
+      Projet.delete_all
+      redirect_to root_path, notice: t('reinitialisation.succes')
+    end
   end
 end
